lm-cli/pyproject.toml
[tool.poetry]
name = "lm-cli"
version = "2.2.12"
description = "License Manager CLI Client"
authors = ["Omnivector Solutions <info@omnivector.solutions>"]
license = "MIT"
packages = [ { include = "lm_cli" } ]


[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.1"
rich = "^12.4.4"
pyperclip = "^1.8.2"
python-jose = "^3.3.0"
loguru = "^0.6.0"
py-buzz = "^3.1.1"
typer = "^0.4.1"
pendulum = "^2.1.2"
httpx = "^0.23.0"
importlib-metadata = "^4.11.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
plummet = "^1.1.0"
pytest = "^7.1.2"
pytest-random-order = "^1.0.4"
pytest-cov = "^3.0.0"
pytest-responsemock = "^1.1.1"
pytest-env = "^0.6.2"
pytest-mock = "^3.8.2"
respx = "^0.19.2"
pyproject-flake8 = "^0.0.1-alpha.4"

[tool.poetry.scripts]
lm-cli = "lm_cli.main:app"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--random-order",
    "--cov=lm_cli",
    "--cov-report=term-missing",
    "--cov-fail-under=75",
]
env = [
    "LM_API_ENDPOINT = https://armada-k8s.staging.omnivector.solutions/lm/api/v1",
    "LM_DEBUG = false",
    "OIDC_DOMAIN = dummy_auth_domain.com",
    "OIDC_AUDIENCE = https://dummy_auth_audience.com",
    "OIDC_CLIENT_ID = dummy_client_id",
    "OIDC_CLIENT_SECRET = dummy_client_secret",
    "IDENTITY_CLAIMS_KEY = https://test-domain.test",
]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
lines_after_imports = "2"
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 120
max-complexity = 40

[[tool.mypy.overrides]]
module = [
    "pyperclip",
    "jose.*",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

